How to create Simple Calculator using Simple HTML Code with JavaScript
Here are the steps to create a simple calculator using HTML and JavaScript which can evaluate simple arithmetic on integer numbers. Two types of inputs text and button are used here on a table within a form element and OnClick event was used to insert button values on the screen or to evaluate the numbers.
 InfoTechSite > Web Designing Tutorials > How to Create a Simple calculator Using HTML and JavaScript
How to Create a Simple calculator Using HTML and JavaScript

shuseel February 23, 2017 0 Comments

Web Designing Tutorials

JavaScript, Programming Tutorials

Here are the steps to create a simple calculator using HTML and JavaScript which can evaluate simple arithmetic on integer numbers. Two types of inputs text and button are used here on a table within a form element and OnClick event was used to insert button values on the screen or to evaluate the numbers.
Steps to create a Simple calculator Using HTML and JavaScript

1. At first Insert a <form> element within <body> tag.
2. Create a table using <table> …..</table> tag.
3. Insert two types of Input text and button within table data of table row using <tr><td>….</td></tr> tag.
4. Assign OnClick event for all the buttons having numbers and arithmetic operators.
5. Give blank value for Clear(C) button.
6. Use eval() function to evaluate the numbers on OnClick event of equal to sign button.